Basic Small-Scale Model of Motor Controller
Support for Bluetooth communication to choose desired mode (forward, reverse, braking, and idle) as well as the target current for the three relevant modes. Includes support for voltage detection for simulating the current sensing module to be used in the future.

Does not yet include critical features such as interrupts for voltage detection routines, PWM for bootstrapping gate drivers, and passing critical sensor data to the parent Raspberry Pi. I think multithreading may also be helpful in reducing latency and conflicts between Bluetooth communication and the voltage sampling routines, but it's necessity has yet to be determined.